Music Hotfix
"Play" function as a command for the bot to join the voice channel of the message author if they are connected; otherwise, it sends a message indicating the requirement to be in a voice channel.
---
"Leave" function as a command for the bot to disconnect from the voice channel it is currently in if connected; otherwise, it sends a message indicating that it is not in a voice channel.

Albo music.py where soon new commands will appear, that provide user to play music on vc